requirements.txt
phidata
sentence-transformers
pydantic
python-dotenv
lancedb
torch
transformers
